requirements.txt
python-dotenv
kopf
pyyaml
kubernetes
requests
